requirements.txt
alabaster==0.7.12
asgiref==3.4.1
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
Django==3.2.9
django-extensions==3.1.5
django-ninja==0.16.1
dnspython==2.1.0
docutils==0.16
ecdsa==0.17.0
email-validator==1.1.3
idna==3.3
imagesize==1.3.0
Jinja2==3.0.3
jose==1.0.0
MarkupSafe==2.0.1
none==0.1.0a2
packaging==21.3
Pillow==8.4.0
pyasn1==0.4.8
pydantic==1.8.2
Pygments==2.10.0
pyparsing==3.0.6
python-jose==3.3.0
pytz==2021.3
requests==2.26.0
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
snowballstemmer==2.2.0
Sphinx==3.5.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-websupport==1.2.4
sqlparse==0.4.2
typing-extensions==3.10.0.2
tzdata==2021.5
urllib3==1.26.7
